src/evaluator.py
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.metrics import precision_recall_curve

class DuplicateEvaluator:
    def __init__(self, similarities):
        self.similarities = similarities
        
    def analyze_distribution(self):
        """Analyze similarity score distribution"""
        plt.figure(figsize=(10, 6))
        sns.histplot(self.similarities, bins=50, kde=True)
        plt.title('Job Description Similarity Distribution')
        plt.xlabel('Cosine Similarity Score')
        plt.ylabel('Count')
        plt.savefig('results/similarity_distribution.png')
        
    def calculate_metrics(self, threshold):
        """Calculate precision/recall metrics"""
        # This would require labeled data for proper evaluation
        # Placeholder for demonstration
        precision = np.mean([s > threshold for s in self.similarities])
        return {'threshold': threshold, 'estimated_precision': precision}

    def find_optimal_threshold(self):
        """Find optimal threshold using elbow method"""
        thresholds = np.linspace(0.5, 0.95, 20)
        densities = [np.mean(self.similarities >= t) for t in thresholds]
        
        plt.plot(thresholds, densities)
        plt.title('Threshold vs. Match Density')
        plt.savefig('results/threshold_analysis.png')